spring-boot-sample-simple: Add enforcer plugin (#18)
Enforce the definition of the property com.redhat.xpaas.repo.redhatag.

This is to aid tests to ensure that the redhatga repo is enabled. We
aren't testing exactly that -- the repo is enabled by the property in
most quickstarts, and this tests the property, not the repo - but it's
close.